US11: Compute relative exposure
- Add relativeExposure method to compute the relative exposure between two images.
- Add computeRelExposure method to compute the relative exposure between each pair of adjacent images in a stack.

+++ MetaData.cpp
@@ -51,7 +51,7 @@ MetaData::MetaData(const char * f, const LibRaw & rawData) : fileName(f) {
 }
 
 
-float MetaData::logExp() const {
+double MetaData::logExp() const {
     return std::log2(isoSpeed * shutter / (100.0 * aperture * aperture));
 }
